requirements.txt
apache-airflow
lxml
pyarrow
pydantic > 2.0
pydantic-xml
pytest
httpx==0.27.0
pandas<2.2.0
paramiko==3.4.0
pendulum==3.0.0
py7zr==0.20.8
pymssql
SQLAlchemy
pymysql
apache-airflow-providers-microsoft-azure
mock
